\name{CH1903}
\alias{CH1903}
\docType{data}
\title{
  \code{CH1903}
}
\description{
  Swiss square projection (CH1903 / LV03)
}
\usage{data(CH1903)}
\format{
  The format is:
Formal class 'CRS' [package "sp"] with 1 slots
  ..@ projargs: chr "+init=epsg:21781 +proj=somerc +lat_0=46.95240555555556 +l on_0=7.439583333333333 
+k_0=1 +x_0=600000 +y_0=200000 +ellps=bessel +t"| __truncated__
}
\keyword{datasets}
\keyword{spatial}
