\name{geospacom-package}
\alias{geospacom-package}
\alias{geospacom}
\docType{package}
\title{
Helper package to facilitate the generation of distance matrices used in
the package \code{\link[spacom:spacom-package]{spacom}}
}
\description{
   Generates distance matrices from shape files and represents spatially
   weighted multilevel analysis results (see \code{spacom}) 
}
\details{
\tabular{ll}{
Package: \tab geospacom\cr
Type: \tab Package\cr
Version: \tab 0.5-4\cr
Date: \tab 2013-03-17\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
  Davide Morselli,
  Mathieu Cossuta,
  Till Junge,
  Sandra Penic,
  Guy Elcheroth,
  

Maintainer: Davide Morselli <davide.morselli@tiscali.it>
}
\keyword{ package }
\keyword{spatial}
\seealso{
\code{\link[spacom:spacom-package]{spacom}}
}
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
