\name{yu.reg}
\alias{yu.reg}
\docType{data}
\title{
  \code{yu.reg}
}
\description{
  A \code{\link[sp]{SpatialPolygonsDataFrame}} shape file of the Former-Yugoslavia territory, divided in 80 polygons
}
\format{
  \code{area.name} - Character variable. Contains names of the 80 polygons
}
 \source{TRACES. Additional data referring to this shapefile can be found in the TRACES datasets,  accessable through the Data and Research Information Services from the Swiss Foundation for Research in the Social Sciences (www.unil.ch/daris).
}

\usage{data(yu.reg)}
\references{
Spini, D., Elcheroth, G., & Fasel, R. (2011). TRACES: Methodological and technical report. \emph{LIVES Working Papers, 4}. DOI: 10.12682/lives.2296-1658.2011.4  
     }
\keyword{datasets}
\keyword{spatial}
